requirements.txt
streamlit
pandas
numpy
joblib
Pillow
requests